Updated maintenanceDuty.py to resolve issue #1
When indexing the "maintenanceDF_yester" dataframe after determining the EM before the day watch was the last day of the previous month, the program attempted to get the EM shift data using the "dateNumeric" variable. This "dateNumeric" variable was not updated to reflect "yesterday's date" which is needed to access the maintenance duty for the EM in this case. I've added a new variable to store "yesterday's date" (called "yesterdayNumeric") which will be used while indexing the maintenance data of the previous month.